requirements.txt
flask==2.3.3
supabase==2.2.1
python-dotenv==1.0.0
apscheduler==3.10.4
gunicon
